+++ linkerd/core/src/main/scala/io/buoyant/linkerd/ClientConfig.scala
@@ -6,6 +6,7 @@ import com.twitter.finagle.Stack
 import com.twitter.finagle.buoyant.{ClientAuth, PathMatcher, TlsClientConfig => FTlsClientConfig}
 import com.twitter.finagle.client.DefaultPool
 import com.twitter.finagle.service._
+import com.twitter.finagle.buoyant.ParamsMaybeWith
 import io.buoyant.router.RetryBudgetConfig
 import io.buoyant.router.RetryBudgetModule.param
