+++ linkerd/core/src/main/scala/io/buoyant/linkerd/package.scala
@@ -1,7 +1,4 @@
 package io.buoyant
-
-import com.twitter.finagle.Stack
-
 /**
  * Linkerd provides a modular & pluggable configuration layer to
  * support programmatic and configuration-driven initialization of
@@ -38,29 +35,4 @@ import com.twitter.finagle.Stack
  * configuration and initialization. ProtocolInitializer modules are
  * discovered at runtime with finagle's `LoadService` facility.
  */
-package object linkerd {
-  implicit class MaybeTransform[A](val a: A) extends AnyVal {
-    def maybeTransform(f: Option[A => A]): A = {
-      f match {
-        case Some(f) => f(a)
-        case None => a
-      }
-    }
-  }
-
-  implicit class ParamsMaybeWith(val params: Stack.Params) extends AnyVal {
-    def maybeWith[T: Stack.Param](p: Option[T]): Stack.Params = {
-      p match {
-        case Some(t) => params + t
-        case None => params
-      }
-    }
-
-    def maybeWith(ps: Option[Stack.Params]): Stack.Params = {
-      ps match {
-        case Some(ps) => params ++ ps
-        case None => params
-      }
-    }
-  }
-}
+package object linkerd
